refactor api certificate validators into common helper
