+++ src/Sonarr.Api.V3/Config/HostConfigResource.cs
@@ -4,10 +4,11 @@ using NzbDrone.Core.Configuration;
 using NzbDrone.Core.Security;
 using NzbDrone.Core.Update;
 using Sonarr.Http.REST;
+using Sonarr.Http.Validation;
 
 namespace Sonarr.Api.V3.Config
 {
-    public class HostConfigResource : RestResource
+    public class HostConfigResource : RestResource, ISslCertificateResource
     {
         public string BindAddress { get; set; }
         public int Port { get; set; }
