+++ src/Sonarr.Http/Validation/ISslCertificateResource.cs
@@ -1,0 +1,8 @@
+namespace Sonarr.Http.Validation;
+
+public interface ISslCertificateResource
+{
+    string SslCertPath { get; }
+    string SslKeyPath { get; }
+    string SslCertPassword { get; }
+}
